+++ README
@@ -1,8 +1,16 @@
 
+Korean romanized input method with hanja words, brute version with a HUGE dictionary.
+
 This input method for RIME is made for my friend, who is a language teacher, speaks a lot of languages.
 
 You can contact me using Chinese/Japanese/English.
 Sorry that I cannot understand Korean without kr-ja Google Translation.
+
+-------------------------
+
+韩文（包括汉字词）罗马音输入法，为每个词列举三类可能的罗马音组合。大词典暴力制作。
+
+此作品初衷为简化一友人之工作，而笔者目前不懂韩文，或多有欠佳处理，但愿不会阻碍您的日常使用。
 
 -------------------------
 Reference:
