--- studio/cb_table.hpp
@@ -1,472 +1,0 @@
-// jamo combination table
-
-#ifndef _CB_TABLE__
-#define _CB_TABLE__
-
-#include <stdint.h>
-
-struct cb_t {
-	uint16_t a;
-	uint16_t b;
-	uint16_t r;
-};
-
-struct jt_t {
-	uint16_t tr;
-	uint16_t rl;
-};
-
-static const uint16_t cb_tbl[] = {
-	0x1100, 0x1100, 0x1101, // ᄀ + ᄀ = ᄁ
-	0x1100, 0x1103, 0x115a, // ᄀ + ᄃ = ᅚ
-	0x1102, 0x1100, 0x1113, // ᄂ + ᄀ = ᄓ
-	0x1102, 0x1102, 0x1114, // ᄂ + ᄂ = ᄔ
-	0x1102, 0x1103, 0x1115, // ᄂ + ᄃ = ᄕ
-	0x1102, 0x1107, 0x1116, // ᄂ + ᄇ = ᄖ
-	0x1102, 0x1109, 0x115b, // ᄂ + ᄉ = ᅛ
-	0x1102, 0x110c, 0x115c, // ᄂ + ᄌ = ᅜ
-	0x1102, 0x1112, 0x115d, // ᄂ + ᄒ = ᅝ
-	0x1103, 0x1100, 0x1117, // ᄃ + ᄀ = ᄗ
-	0x1103, 0x1103, 0x1104, // ᄃ + ᄃ = ᄄ
-	0x1103, 0x1105, 0x115e, // ᄃ + ᄅ = ᅞ
-	0x1103, 0x1106, 0xa960, // ᄃ + ᄆ = ꥠ
-	0x1103, 0x1107, 0xa961, // ᄃ + ᄇ = ꥡ
-	0x1103, 0x1109, 0xa962, // ᄃ + ᄉ = ꥢ
-	0x1103, 0x110c, 0xa963, // ᄃ + ᄌ = ꥣ
-	0x1105, 0x1100, 0xa964, // ᄅ + ᄀ = ꥤ
-	0x1105, 0x1101, 0xa965, // ᄅ + ᄁ = ꥥ
-	0x1105, 0x1102, 0x1118, // ᄅ + ᄂ = ᄘ
-	0x1105, 0x1103, 0xa966, // ᄅ + ᄃ = ꥦ
-	0x1105, 0x1104, 0xa967, // ᄅ + ᄄ = ꥧ
-	0x1105, 0x1105, 0x1119, // ᄅ + ᄅ = ᄙ
-	0x1105, 0x1106, 0xa968, // ᄅ + ᄆ = ꥨ
-	0x1105, 0x1107, 0xa969, // ᄅ + ᄇ = ꥩ
-	0x1105, 0x1108, 0xa96a, // ᄅ + ᄈ = ꥪ
-	0x1105, 0x1109, 0xa96c, // ᄅ + ᄉ = ꥬ
-	0x1105, 0x110b, 0x111b, // ᄅ + ᄋ = ᄛ
-	0x1105, 0x110c, 0xa96d, // ᄅ + ᄌ = ꥭ
-	0x1105, 0x110f, 0xa96e, // ᄅ + ᄏ = ꥮ
-	0x1105, 0x1112, 0x111a, // ᄅ + ᄒ = ᄚ
-	0x1105, 0x112b, 0xa96b, // ᄅ + ᄫ = ꥫ
-	0x1106, 0x1100, 0xa96f, // ᄆ + ᄀ = ꥯ
-	0x1106, 0x1103, 0xa970, // ᄆ + ᄃ = ꥰ
-	0x1106, 0x1107, 0x111c, // ᄆ + ᄇ = ᄜ
-	0x1106, 0x1109, 0xa971, // ᄆ + ᄉ = ꥱ
-	0x1106, 0x110b, 0x111d, // ᄆ + ᄋ = ᄝ
-	0x1107, 0x1100, 0x111e, // ᄇ + ᄀ = ᄞ
-	0x1107, 0x1102, 0x111f, // ᄇ + ᄂ = ᄟ
-	0x1107, 0x1103, 0x1120, // ᄇ + ᄃ = ᄠ
-	0x1107, 0x1107, 0x1108, // ᄇ + ᄇ = ᄈ
-	0x1107, 0x1109, 0x1121, // ᄇ + ᄉ = ᄡ
-	0x1107, 0x110a, 0x1125, // ᄇ + ᄊ = ᄥ
-	0x1107, 0x110b, 0x112b, // ᄇ + ᄋ = ᄫ
-	0x1107, 0x110c, 0x1127, // ᄇ + ᄌ = ᄧ
-	0x1107, 0x110e, 0x1128, // ᄇ + ᄎ = ᄨ
-	0x1107, 0x110f, 0xa973, // ᄇ + ᄏ = ꥳ
-	0x1107, 0x1110, 0x1129, // ᄇ + ᄐ = ᄩ
-	0x1107, 0x1111, 0x112a, // ᄇ + ᄑ = ᄪ
-	0x1107, 0x1112, 0xa974, // ᄇ + ᄒ = ꥴ
-	0x1107, 0x112b, 0x112c, // ᄇ + ᄫ = ᄬ
-	0x1107, 0x112d, 0x1122, // ᄇ + ᄭ = ᄢ
-	0x1107, 0x112f, 0x1123, // ᄇ + ᄯ = ᄣ
-	0x1107, 0x1132, 0x1124, // ᄇ + ᄲ = ᄤ
-	0x1107, 0x1136, 0x1126, // ᄇ + ᄶ = ᄦ
-	0x1107, 0x1139, 0xa972, // ᄇ + ᄹ = ꥲ
-	0x1108, 0x110b, 0x112c, // ᄈ + ᄋ = ᄬ
-	0x1109, 0x1100, 0x112d, // ᄉ + ᄀ = ᄭ
-	0x1109, 0x1102, 0x112e, // ᄉ + ᄂ = ᄮ
-	0x1109, 0x1103, 0x112f, // ᄉ + ᄃ = ᄯ
-	0x1109, 0x1105, 0x1130, // ᄉ + ᄅ = ᄰ
-	0x1109, 0x1106, 0x1131, // ᄉ + ᄆ = ᄱ
-	0x1109, 0x1107, 0x1132, // ᄉ + ᄇ = ᄲ
-	0x1109, 0x1109, 0x110a, // ᄉ + ᄉ = ᄊ
-	0x1109, 0x110a, 0x1134, // ᄉ + ᄊ = ᄴ
-	0x1109, 0x110b, 0x1135, // ᄉ + ᄋ = ᄵ
-	0x1109, 0x110c, 0x1136, // ᄉ + ᄌ = ᄶ
-	0x1109, 0x110e, 0x1137, // ᄉ + ᄎ = ᄷ
-	0x1109, 0x110f, 0x1138, // ᄉ + ᄏ = ᄸ
-	0x1109, 0x1110, 0x1139, // ᄉ + ᄐ = ᄹ
-	0x1109, 0x1111, 0x113a, // ᄉ + ᄑ = ᄺ
-	0x1109, 0x1112, 0x113b, // ᄉ + ᄒ = ᄻ
-	0x1109, 0x111e, 0x1133, // ᄉ + ᄞ = ᄳ
-	0x1109, 0x1132, 0xa975, // ᄉ + ᄲ = ꥵ
-	0x110a, 0x1107, 0xa975, // ᄊ + ᄇ = ꥵ
-	0x110a, 0x1109, 0x1134, // ᄊ + ᄉ = ᄴ
-	0x110b, 0x1100, 0x1141, // ᄋ + ᄀ = ᅁ
-	0x110b, 0x1103, 0x1142, // ᄋ + ᄃ = ᅂ
-	0x110b, 0x1105, 0xa976, // ᄋ + ᄅ = ꥶ
-	0x110b, 0x1106, 0x1143, // ᄋ + ᄆ = ᅃ
-	0x110b, 0x1107, 0x1144, // ᄋ + ᄇ = ᅄ
-	0x110b, 0x1109, 0x1145, // ᄋ + ᄉ = ᅅ
-	0x110b, 0x110b, 0x1147, // ᄋ + ᄋ = ᅇ
-	0x110b, 0x110c, 0x1148, // ᄋ + ᄌ = ᅈ
-	0x110b, 0x110e, 0x1149, // ᄋ + ᄎ = ᅉ
-	0x110b, 0x1110, 0x114a, // ᄋ + ᄐ = ᅊ
-	0x110b, 0x1111, 0x114b, // ᄋ + ᄑ = ᅋ
-	0x110b, 0x1112, 0xa977, // ᄋ + ᄒ = ꥷ
-	0x110b, 0x1140, 0x1146, // ᄋ + ᅀ = ᅆ
-	0x110c, 0x110b, 0x114d, // ᄌ + ᄋ = ᅍ
-	0x110c, 0x110c, 0x110d, // ᄌ + ᄌ = ᄍ
-	0x110d, 0x1112, 0xa978, // ᄍ + ᄒ = ꥸ
-	0x110e, 0x110f, 0x1152, // ᄎ + ᄏ = ᅒ
-	0x110e, 0x1112, 0x1153, // ᄎ + ᄒ = ᅓ
-	0x1110, 0x1110, 0xa979, // ᄐ + ᄐ = ꥹ
-	0x1111, 0x1107, 0x1156, // ᄑ + ᄇ = ᅖ
-	0x1111, 0x110b, 0x1157, // ᄑ + ᄋ = ᅗ
-	0x1111, 0x1112, 0xa97a, // ᄑ + ᄒ = ꥺ
-	0x1112, 0x1109, 0xa97b, // ᄒ + ᄉ = ꥻ
-	0x1112, 0x1112, 0x1158, // ᄒ + ᄒ = ᅘ
-	0x1121, 0x1100, 0x1122, // ᄡ + ᄀ = ᄢ
-	0x1121, 0x1103, 0x1123, // ᄡ + ᄃ = ᄣ
-	0x1121, 0x1107, 0x1124, // ᄡ + ᄇ = ᄤ
-	0x1121, 0x1109, 0x1125, // ᄡ + ᄉ = ᄥ
-	0x1121, 0x110c, 0x1126, // ᄡ + ᄌ = ᄦ
-	0x1121, 0x1110, 0xa972, // ᄡ + ᄐ = ꥲ
-	0x1132, 0x1100, 0x1133, // ᄲ + ᄀ = ᄳ
-	0x113c, 0x113c, 0x113d, // ᄼ + ᄼ = ᄽ
-	0x113e, 0x113e, 0x113f, // ᄾ + ᄾ = ᄿ
-	0x114e, 0x114e, 0x114f, // ᅎ + ᅎ = ᅏ
-	0x1150, 0x1150, 0x1151, // ᅐ + ᅐ = ᅑ
-	0x1159, 0x1159, 0xa97c, // ᅙ + ᅙ = ꥼ
-	0x1161, 0x1161, 0x119e, // ᅡ + ᅡ = ᆞ
-	0x1161, 0x1169, 0x1176, // ᅡ + ᅩ = ᅶ
-	0x1161, 0x116e, 0x1177, // ᅡ + ᅮ = ᅷ
-	0x1161, 0x1173, 0x11a3, // ᅡ + ᅳ = ᆣ
-	0x1161, 0x1175, 0x1162, // ᅡ + ᅵ = ᅢ
-	0x1163, 0x1169, 0x1178, // ᅣ + ᅩ = ᅸ
-	0x1163, 0x116d, 0x1179, // ᅣ + ᅭ = ᅹ
-	0x1163, 0x116e, 0x11a4, // ᅣ + ᅮ = ᆤ
-	0x1163, 0x1175, 0x1164, // ᅣ + ᅵ = ᅤ
-	0x1165, 0x1169, 0x117a, // ᅥ + ᅩ = ᅺ
-	0x1165, 0x116e, 0x117b, // ᅥ + ᅮ = ᅻ
-	0x1165, 0x1173, 0x117c, // ᅥ + ᅳ = ᅼ
-	0x1165, 0x1175, 0x1166, // ᅥ + ᅵ = ᅦ
-	0x1167, 0x1163, 0x11a5, // ᅧ + ᅣ = ᆥ
-	0x1167, 0x1169, 0x117d, // ᅧ + ᅩ = ᅽ
-	0x1167, 0x116e, 0x117e, // ᅧ + ᅮ = ᅾ
-	0x1167, 0x1175, 0x1168, // ᅧ + ᅵ = ᅨ
-	0x1169, 0x1161, 0x116a, // ᅩ + ᅡ = ᅪ
-	0x1169, 0x1162, 0x116b, // ᅩ + ᅢ = ᅫ
-	0x1169, 0x1163, 0x11a6, // ᅩ + ᅣ = ᆦ
-	0x1169, 0x1164, 0x11a7, // ᅩ + ᅤ = ᆧ
-	0x1169, 0x1165, 0x117f, // ᅩ + ᅥ = ᅿ
-	0x1169, 0x1166, 0x1180, // ᅩ + ᅦ = ᆀ
-	0x1169, 0x1167, 0xd7b0, // ᅩ + ᅧ = ힰ
-	0x1169, 0x1168, 0x1181, // ᅩ + ᅨ = ᆁ
-	0x1169, 0x1169, 0x1182, // ᅩ + ᅩ = ᆂ
-	0x1169, 0x116e, 0x1183, // ᅩ + ᅮ = ᆃ
-	0x1169, 0x1175, 0x116c, // ᅩ + ᅵ = ᅬ
-	0x116a, 0x1175, 0x116b, // ᅪ + ᅵ = ᅫ
-	0x116d, 0x1161, 0xd7b2, // ᅭ + ᅡ = ힲ
-	0x116d, 0x1162, 0xd7b3, // ᅭ + ᅢ = ힳ
-	0x116d, 0x1163, 0x1184, // ᅭ + ᅣ = ᆄ
-	0x116d, 0x1164, 0x1185, // ᅭ + ᅤ = ᆅ
-	0x116d, 0x1165, 0xd7b4, // ᅭ + ᅥ = ힴ
-	0x116d, 0x1167, 0x1186, // ᅭ + ᅧ = ᆆ
-	0x116d, 0x1169, 0x1187, // ᅭ + ᅩ = ᆇ
-	0x116d, 0x1175, 0x1188, // ᅭ + ᅵ = ᆈ
-	0x116e, 0x1161, 0x1189, // ᅮ + ᅡ = ᆉ
-	0x116e, 0x1162, 0x118a, // ᅮ + ᅢ = ᆊ
-	0x116e, 0x1165, 0x116f, // ᅮ + ᅥ = ᅯ
-	0x116e, 0x1166, 0x1170, // ᅮ + ᅦ = ᅰ
-	0x116e, 0x1167, 0xd7b5, // ᅮ + ᅧ = ힵ
-	0x116e, 0x1168, 0x118c, // ᅮ + ᅨ = ᆌ
-	0x116e, 0x116e, 0x118d, // ᅮ + ᅮ = ᆍ
-	0x116e, 0x1175, 0x1171, // ᅮ + ᅵ = ᅱ
-	0x116e, 0x117c, 0x118b, // ᅮ + ᅼ = ᆋ
-	0x116e, 0xd7c4, 0xd7b6, // ᅮ + ퟄ = ힶ
-	0x116f, 0x1173, 0x118b, // ᅯ + ᅳ = ᆋ
-	0x116f, 0x1175, 0x1170, // ᅯ + ᅵ = ᅰ
-	0x1171, 0x1175, 0xd7b6, // ᅱ + ᅵ = ힶ
-	0x1172, 0x1161, 0x118e, // ᅲ + ᅡ = ᆎ
-	0x1172, 0x1162, 0xd7b7, // ᅲ + ᅢ = ힷ
-	0x1172, 0x1165, 0x118f, // ᅲ + ᅥ = ᆏ
-	0x1172, 0x1166, 0x1190, // ᅲ + ᅦ = ᆐ
-	0x1172, 0x1167, 0x1191, // ᅲ + ᅧ = ᆑ
-	0x1172, 0x1168, 0x1192, // ᅲ + ᅨ = ᆒ
-	0x1172, 0x1169, 0xd7b8, // ᅲ + ᅩ = ힸ
-	0x1172, 0x116e, 0x1193, // ᅲ + ᅮ = ᆓ
-	0x1172, 0x1175, 0x1194, // ᅲ + ᅵ = ᆔ
-	0x1173, 0x1161, 0xd7b9, // ᅳ + ᅡ = ힹ
-	0x1173, 0x1165, 0xd7ba, // ᅳ + ᅥ = ힺ
-	0x1173, 0x1166, 0xd7bb, // ᅳ + ᅦ = ힻ
-	0x1173, 0x1169, 0xd7bc, // ᅳ + ᅩ = ힼ
-	0x1173, 0x116e, 0x1195, // ᅳ + ᅮ = ᆕ
-	0x1173, 0x1173, 0x1196, // ᅳ + ᅳ = ᆖ
-	0x1173, 0x1175, 0x1174, // ᅳ + ᅵ = ᅴ
-	0x1174, 0x116e, 0x1197, // ᅴ + ᅮ = ᆗ
-	0x1175, 0x1161, 0x1198, // ᅵ + ᅡ = ᆘ
-	0x1175, 0x1163, 0x1199, // ᅵ + ᅣ = ᆙ
-	0x1175, 0x1164, 0xd7be, // ᅵ + ᅤ = ힾ
-	0x1175, 0x1167, 0xd7bf, // ᅵ + ᅧ = ힿ
-	0x1175, 0x1168, 0xd7c0, // ᅵ + ᅨ = ퟀ
-	0x1175, 0x1169, 0x119a, // ᅵ + ᅩ = ᆚ
-	0x1175, 0x116d, 0xd7c2, // ᅵ + ᅭ = ퟂ
-	0x1175, 0x116e, 0x119b, // ᅵ + ᅮ = ᆛ
-	0x1175, 0x1172, 0xd7c3, // ᅵ + ᅲ = ퟃ
-	0x1175, 0x1173, 0x119c, // ᅵ + ᅳ = ᆜ
-	0x1175, 0x1175, 0xd7c4, // ᅵ + ᅵ = ퟄ
-	0x1175, 0x1178, 0xd7bd, // ᅵ + ᅸ = ힽ
-	0x1175, 0x119e, 0x119d, // ᅵ + ᆞ = ᆝ
-	0x1182, 0x1175, 0xd7b1, // ᆂ + ᅵ = ힱ
-	0x1199, 0x1169, 0xd7bd, // ᆙ + ᅩ = ힽ
-	0x119a, 0x1175, 0xd7c1, // ᆚ + ᅵ = ퟁ
-	0x119e, 0x1161, 0xd7c5, // ᆞ + ᅡ = ퟅ
-	0x119e, 0x1165, 0x119f, // ᆞ + ᅥ = ᆟ
-	0x119e, 0x1166, 0xd7c6, // ᆞ + ᅦ = ퟆ
-	0x119e, 0x116e, 0x11a0, // ᆞ + ᅮ = ᆠ
-	0x119e, 0x1175, 0x11a1, // ᆞ + ᅵ = ᆡ
-	0x119e, 0x119e, 0x11a2, // ᆞ + ᆞ = ᆢ
-	0x11a8, 0x11a8, 0x11a9, // ᆨ + ᆨ = ᆩ
-	0x11a8, 0x11ab, 0x11fa, // ᆨ + ᆫ = ᇺ
-	0x11a8, 0x11af, 0x11c3, // ᆨ + ᆯ = ᇃ
-	0x11a8, 0x11b8, 0x11fb, // ᆨ + ᆸ = ᇻ
-	0x11a8, 0x11ba, 0x11aa, // ᆨ + ᆺ = ᆪ
-	0x11a8, 0x11be, 0x11fc, // ᆨ + ᆾ = ᇼ
-	0x11a8, 0x11bf, 0x11fd, // ᆨ + ᆿ = ᇽ
-	0x11a8, 0x11c2, 0x11fe, // ᆨ + ᇂ = ᇾ
-	0x11a8, 0x11e7, 0x11c4, // ᆨ + ᇧ = ᇄ
-	0x11aa, 0x11a8, 0x11c4, // ᆪ + ᆨ = ᇄ
-	0x11ab, 0x11a8, 0x11c5, // ᆫ + ᆨ = ᇅ
-	0x11ab, 0x11ab, 0x11ff, // ᆫ + ᆫ = ᇿ
-	0x11ab, 0x11ae, 0x11c6, // ᆫ + ᆮ = ᇆ
-	0x11ab, 0x11af, 0xd7cb, // ᆫ + ᆯ = ퟋ
-	0x11ab, 0x11ba, 0x11c7, // ᆫ + ᆺ = ᇇ
-	0x11ab, 0x11bd, 0x11ac, // ᆫ + ᆽ = ᆬ
-	0x11ab, 0x11be, 0xd7cc, // ᆫ + ᆾ = ퟌ
-	0x11ab, 0x11c0, 0x11c9, // ᆫ + ᇀ = ᇉ
-	0x11ab, 0x11c2, 0x11ad, // ᆫ + ᇂ = ᆭ
-	0x11ab, 0x11eb, 0x11c8, // ᆫ + ᇫ = ᇈ
-	0x11ae, 0x11a8, 0x11ca, // ᆮ + ᆨ = ᇊ
-	0x11ae, 0x11ae, 0xd7cd, // ᆮ + ᆮ = ퟍ
-	0x11ae, 0x11af, 0x11cb, // ᆮ + ᆯ = ᇋ
-	0x11ae, 0x11b8, 0xd7cf, // ᆮ + ᆸ = ퟏ
-	0x11ae, 0x11ba, 0xd7d0, // ᆮ + ᆺ = ퟐ
-	0x11ae, 0x11bd, 0xd7d2, // ᆮ + ᆽ = ퟒ
-	0x11ae, 0x11be, 0xd7d3, // ᆮ + ᆾ = ퟓ
-	0x11ae, 0x11c0, 0xd7d4, // ᆮ + ᇀ = ퟔ
-	0x11ae, 0x11e7, 0xd7d1, // ᆮ + ᇧ = ퟑ
-	0x11ae, 0xd7cf, 0xd7ce, // ᆮ + ퟏ = ퟎ
-	0x11af, 0x11a8, 0x11b0, // ᆯ + ᆨ = ᆰ
-	0x11af, 0x11a9, 0xd7d5, // ᆯ + ᆩ = ퟕ
-	0x11af, 0x11aa, 0x11cc, // ᆯ + ᆪ = ᇌ
-	0x11af, 0x11ab, 0x11cd, // ᆯ + ᆫ = ᇍ
-	0x11af, 0x11ae, 0x11ce, // ᆯ + ᆮ = ᇎ
-	0x11af, 0x11af, 0x11d0, // ᆯ + ᆯ = ᇐ
-	0x11af, 0x11b7, 0x11b1, // ᆯ + ᆷ = ᆱ
-	0x11af, 0x11b8, 0x11b2, // ᆯ + ᆸ = ᆲ
-	0x11af, 0x11b9, 0x11d3, // ᆯ + ᆹ = ᇓ
-	0x11af, 0x11ba, 0x11b3, // ᆯ + ᆺ = ᆳ
-	0x11af, 0x11bb, 0x11d6, // ᆯ + ᆻ = ᇖ
-	0x11af, 0x11bc, 0xd7dd, // ᆯ + ᆼ = ퟝ
-	0x11af, 0x11bf, 0x11d8, // ᆯ + ᆿ = ᇘ
-	0x11af, 0x11c0, 0x11b4, // ᆯ + ᇀ = ᆴ
-	0x11af, 0x11c1, 0x11b5, // ᆯ + ᇁ = ᆵ
-	0x11af, 0x11c2, 0x11b6, // ᆯ + ᇂ = ᆶ
-	0x11af, 0x11d8, 0xd7d7, // ᆯ + ᇘ = ퟗ
-	0x11af, 0x11da, 0x11d1, // ᆯ + ᇚ = ᇑ
-	0x11af, 0x11dd, 0x11d2, // ᆯ + ᇝ = ᇒ
-	0x11af, 0x11e1, 0xd7d8, // ᆯ + ᇡ = ퟘ
-	0x11af, 0x11e4, 0xd7da, // ᆯ + ᇤ = ퟚ
-	0x11af, 0x11e5, 0x11d4, // ᆯ + ᇥ = ᇔ
-	0x11af, 0x11e6, 0x11d5, // ᆯ + ᇦ = ᇕ
-	0x11af, 0x11eb, 0x11d7, // ᆯ + ᇫ = ᇗ
-	0x11af, 0x11f0, 0xd7db, // ᆯ + ᇰ = ퟛ
-	0x11af, 0x11f9, 0x11d9, // ᆯ + ᇹ = ᇙ
-	0x11af, 0x11fe, 0xd7d6, // ᆯ + ᇾ = ퟖ
-	0x11af, 0xd7e3, 0xd7d9, // ᆯ + ퟣ = ퟙ
-	0x11b0, 0x11a8, 0xd7d5, // ᆰ + ᆨ = ퟕ
-	0x11b0, 0x11ba, 0x11cc, // ᆰ + ᆺ = ᇌ
-	0x11b0, 0x11c2, 0xd7d6, // ᆰ + ᇂ = ퟖ
-	0x11b1, 0x11a8, 0x11d1, // ᆱ + ᆨ = ᇑ
-	0x11b1, 0x11ba, 0x11d2, // ᆱ + ᆺ = ᇒ
-	0x11b1, 0x11c2, 0xd7d8, // ᆱ + ᇂ = ퟘ
-	0x11b2, 0x11ae, 0xd7d9, // ᆲ + ᆮ = ퟙ
-	0x11b2, 0x11ba, 0x11d3, // ᆲ + ᆺ = ᇓ
-	0x11b2, 0x11bc, 0x11d5, // ᆲ + ᆼ = ᇕ
-	0x11b2, 0x11c1, 0xd7da, // ᆲ + ᇁ = ퟚ
-	0x11b2, 0x11c2, 0x11d4, // ᆲ + ᇂ = ᇔ
-	0x11b3, 0x11ba, 0x11d6, // ᆳ + ᆺ = ᇖ
-	0x11b7, 0x11a8, 0x11da, // ᆷ + ᆨ = ᇚ
-	0x11b7, 0x11ab, 0xd7de, // ᆷ + ᆫ = ퟞ
-	0x11b7, 0x11af, 0x11db, // ᆷ + ᆯ = ᇛ
-	0x11b7, 0x11b7, 0xd7e0, // ᆷ + ᆷ = ퟠ
-	0x11b7, 0x11b8, 0x11dc, // ᆷ + ᆸ = ᇜ
-	0x11b7, 0x11b9, 0xd7e1, // ᆷ + ᆹ = ퟡ
-	0x11b7, 0x11ba, 0x11dd, // ᆷ + ᆺ = ᇝ
-	0x11b7, 0x11bb, 0x11de, // ᆷ + ᆻ = ᇞ
-	0x11b7, 0x11bc, 0x11e2, // ᆷ + ᆼ = ᇢ
-	0x11b7, 0x11bd, 0xd7e2, // ᆷ + ᆽ = ퟢ
-	0x11b7, 0x11be, 0x11e0, // ᆷ + ᆾ = ᇠ
-	0x11b7, 0x11c2, 0x11e1, // ᆷ + ᇂ = ᇡ
-	0x11b7, 0x11eb, 0x11df, // ᆷ + ᇫ = ᇟ
-	0x11b7, 0x11ff, 0xd7df, // ᆷ + ᇿ = ퟟ
-	0x11b8, 0x11ae, 0xd7e3, // ᆸ + ᆮ = ퟣ
-	0x11b8, 0x11af, 0x11e3, // ᆸ + ᆯ = ᇣ
-	0x11b8, 0x11b5, 0xd7e4, // ᆸ + ᆵ = ퟤ
-	0x11b8, 0x11b7, 0xd7e5, // ᆸ + ᆷ = ퟥ
-	0x11b8, 0x11b8, 0xd7e6, // ᆸ + ᆸ = ퟦ
-	0x11b8, 0x11ba, 0x11b9, // ᆸ + ᆺ = ᆹ
-	0x11b8, 0x11bc, 0x11e6, // ᆸ + ᆼ = ᇦ
-	0x11b8, 0x11bd, 0xd7e8, // ᆸ + ᆽ = ퟨ
-	0x11b8, 0x11be, 0xd7e9, // ᆸ + ᆾ = ퟩ
-	0x11b8, 0x11c1, 0x11e4, // ᆸ + ᇁ = ᇤ
-	0x11b8, 0x11c2, 0x11e5, // ᆸ + ᇂ = ᇥ
-	0x11b8, 0x11e8, 0xd7e7, // ᆸ + ᇨ = ퟧ
-	0x11b9, 0x11ae, 0xd7e7, // ᆹ + ᆮ = ퟧ
-	0x11ba, 0x11a8, 0x11e7, // ᆺ + ᆨ = ᇧ
-	0x11ba, 0x11ae, 0x11e8, // ᆺ + ᆮ = ᇨ
-	0x11ba, 0x11af, 0x11e9, // ᆺ + ᆯ = ᇩ
-	0x11ba, 0x11b7, 0xd7ea, // ᆺ + ᆷ = ퟪ
-	0x11ba, 0x11b8, 0x11ea, // ᆺ + ᆸ = ᇪ
-	0x11ba, 0x11ba, 0x11bb, // ᆺ + ᆺ = ᆻ
-	0x11ba, 0x11bd, 0xd7ef, // ᆺ + ᆽ = ퟯ
-	0x11ba, 0x11be, 0xd7f0, // ᆺ + ᆾ = ퟰ
-	0x11ba, 0x11c0, 0xd7f1, // ᆺ + ᇀ = ퟱ
-	0x11ba, 0x11c2, 0xd7f2, // ᆺ + ᇂ = ퟲ
-	0x11ba, 0x11e6, 0xd7eb, // ᆺ + ᇦ = ퟫ
-	0x11ba, 0x11e7, 0xd7ec, // ᆺ + ᇧ = ퟬ
-	0x11ba, 0x11e8, 0xd7ed, // ᆺ + ᇨ = ퟭ
-	0x11ba, 0x11eb, 0xd7ee, // ᆺ + ᇫ = ퟮ
-	0x11bb, 0x11a8, 0xd7ec, // ᆻ + ᆨ = ퟬ
-	0x11bb, 0x11ae, 0xd7ed, // ᆻ + ᆮ = ퟭ
-	0x11bd, 0x11b8, 0xd7f7, // ᆽ + ᆸ = ퟷ
-	0x11bd, 0x11bd, 0xd7f9, // ᆽ + ᆽ = ퟹ
-	0x11bd, 0xd7e6, 0xd7f8, // ᆽ + ퟦ = ퟸ
-	0x11c1, 0x11b8, 0x11f3, // ᇁ + ᆸ = ᇳ
-	0x11c1, 0x11ba, 0xd7fa, // ᇁ + ᆺ = ퟺ
-	0x11c1, 0x11bc, 0x11f4, // ᇁ + ᆼ = ᇴ
-	0x11c1, 0x11c0, 0xd7fb, // ᇁ + ᇀ = ퟻ
-	0x11c2, 0x11ab, 0x11f5, // ᇂ + ᆫ = ᇵ
-	0x11c2, 0x11af, 0x11f6, // ᇂ + ᆯ = ᇶ
-	0x11c2, 0x11b7, 0x11f7, // ᇂ + ᆷ = ᇷ
-	0x11c2, 0x11b8, 0x11f8, // ᇂ + ᆸ = ᇸ
-	0x11ce, 0x11c2, 0x11cf, // ᇎ + ᇂ = ᇏ
-	0x11d0, 0x11bf, 0xd7d7, // ᇐ + ᆿ = ퟗ
-	0x11d9, 0x11c2, 0xd7dc, // ᇙ + ᇂ = ퟜ
-	0x11dc, 0x11ba, 0xd7e1, // ᇜ + ᆺ = ퟡ
-	0x11dd, 0x11ba, 0x11de, // ᇝ + ᆺ = ᇞ
-	0x11e3, 0x11c1, 0xd7e4, // ᇣ + ᇁ = ퟤ
-	0x11ea, 0x11bc, 0xd7eb, // ᇪ + ᆼ = ퟫ
-	0x11eb, 0x11b8, 0xd7f3, // ᇫ + ᆸ = ퟳ
-	0x11eb, 0x11e6, 0xd7f4, // ᇫ + ᇦ = ퟴ
-	0x11ec, 0x11a8, 0x11ed, // ᇬ + ᆨ = ᇭ
-	0x11f0, 0x11a8, 0x11ec, // ᇰ + ᆨ = ᇬ
-	0x11f0, 0x11a9, 0x11ed, // ᇰ + ᆩ = ᇭ
-	0x11f0, 0x11b7, 0xd7f5, // ᇰ + ᆷ = ퟵ
-	0x11f0, 0x11ba, 0x11f1, // ᇰ + ᆺ = ᇱ
-	0x11f0, 0x11bf, 0x11ef, // ᇰ + ᆿ = ᇯ
-	0x11f0, 0x11c2, 0xd7f6, // ᇰ + ᇂ = ퟶ
-	0x11f0, 0x11eb, 0x11f2, // ᇰ + ᇫ = ᇲ
-	0x11f0, 0x11f0, 0x11ee, // ᇰ + ᇰ = ᇮ
-	0xa964, 0x1100, 0xa965, // ꥤ + ᄀ = ꥥ
-	0xa966, 0x1103, 0xa967, // ꥦ + ᄃ = ꥧ
-	0xa969, 0x1107, 0xa96a, // ꥩ + ᄇ = ꥪ
-	0xa969, 0x110b, 0xa96b, // ꥩ + ᄋ = ꥫ
-	0xd7c5, 0x1161, 0x11a2, // ퟅ + ᅡ = ᆢ
-	0xd7cd, 0x11b8, 0xd7ce, // ퟍ + ᆸ = ퟎ
-	0xd7d0, 0x11a8, 0xd7d1, // ퟐ + ᆨ = ퟑ
-	0xd7de, 0x11ab, 0xd7df, // ퟞ + ᆫ = ퟟ
-	0xd7f3, 0x11bc, 0xd7f4, // ퟳ + ᆼ = ퟴ
-	0xd7f7, 0x11b8, 0xd7f8, // ퟷ + ᆸ = ퟸ
-	0,  0,  0,
-};
-
-static const uint16_t jamo_trans_tbl[] = {
-	0x3131, 0x1100,
-	0x3132, 0x1101,
-	0x3133, 0x11aa,
-	0x3134, 0x1102,
-	0x3135, 0x11ac,
-	0x3136, 0x11ad,
-	0x3137, 0x1103,
-	0x3138, 0x1104,
-	0x3139, 0x1105,
-	0x313a, 0x11b0,
-	0x313b, 0x11b1,
-	0x313c, 0x11b2,
-	0x313d, 0x11b3,
-	0x313e, 0x11b4,
-	0x313f, 0x11b5,
-	0x3140, 0x11b6,
-	0x3141, 0x1106,
-	0x3142, 0x1107,
-	0x3143, 0x1108,
-	0x3144, 0x11b9,
-	0x3145, 0x1109,
-	0x3146, 0x110a,
-	0x3147, 0x110b,
-	0x3148, 0x110c,
-	0x3149, 0x110d,
-	0x314a, 0x110e,
-	0x314b, 0x110f,
-	0x314c, 0x1110,
-	0x314d, 0x1111,
-	0x314e, 0x1112,
-	0x314f, 0x1161,
-	0x3150, 0x1162,
-	0x3151, 0x1163,
-	0x3152, 0x1164,
-	0x3153, 0x1165,
-	0x3154, 0x1166,
-	0x3155, 0x1167,
-	0x3156, 0x1168,
-	0x3157, 0x1169,
-	0x3158, 0x116a,
-	0x3159, 0x116b,
-	0x315a, 0x116c,
-	0x315b, 0x116d,
-	0x315c, 0x116e,
-	0x315d, 0x116f,
-	0x315e, 0x1170,
-	0x315f, 0x1171,
-	0x3160, 0x1172,
-	0x3161, 0x1173,
-	0x3162, 0x1174,
-	0x3163, 0x1175,
-	0x3164, 0x0000, // ?
-	0x3165, 0x1114, // ?
-	0x3166, 0x1115, // ?
-	0x3167, 0x11c7, // ?
-	0x3168, 0x11c8, // ?
-	0x3169, 0x11cc, // ?
-	0x316a, 0x11ce, // ?
-	0x316b, 0x11d3, // ?
-	0x316c, 0x11d7, // ?
-	0x316d, 0x11d9, // ?
-	0x316e, 0x11dc, // ?
-	0x316f, 0x11dd, // ?
-	0x3170, 0x11df, // ?
-	0x3171, 0x111d, // ?
-	0x3172, 0x111e, // ?
-	0x3173, 0x1120, // ?
-	0x3174, 0x1122, // ?
-	0x3175, 0x1123, // ?
-	0x3176, 0x1127, // ?
-	0x3177, 0x1129, // ?
-	0x3178, 0x112b, // ?
-	0x3179, 0x112c, // ?
-	0x317a, 0x112d, // ?
-	0x317b, 0x112e, // ?
-	0x317c, 0x112f, // ?
-	0x317d, 0x1132, // ?
-	0x317e, 0x1136, // ?
-	0x317f, 0x1140,
-	0x3180, 0x1147, // ?
-	0x3181, 0x114c,
-	0x3182, 0x1145, // ?
-	0x3183, 0x1146, // ?
-	0x3184, 0x1157, // ?
-	0x3185, 0x1158, // ?
-	0x3186, 0x1159,
-	0x3187, 0x1184, // ?
-	0x3188, 0x1185, // ?
-	0x3189, 0x1188, // ?
-	0x318a, 0x1191, // ?
-	0x318b, 0x1192, // ?
-	0x318c, 0x1194, // ?
-	0x318d, 0x119e,
-	0x318e, 0x11a1, // ?
-	0, 0,
-};
-#endif //_CB_TABLE__
